+++ doc/release-notes.md
@@ -71,6 +71,24 @@ Low-level RPC changes
   * `getmininginfo`
 - The wallet RPC `getreceivedbyaddress` will return an error if called with an address not in the wallet.
 
+External wallet files
+---------------------
+
+- The `-wallet=<path>` option now accepts full paths instead of requiring
+  wallets to be located in the bitcoin data directory. This change is backwards
+  compatible since relative paths are interpreted from the bitcoin data
+  directory.
+
+- If `-wallet=<path>` is specified with a path that does not exist, it will now
+  create a wallet directory at the specified location (containing a wallet.dat
+  data file, a db.log file, and database/log.?????????? files) instead of just
+  creating a data file at the path and storing log files in the parent
+  directory. This should make backing up wallets more straightforward than
+  before because the specified wallet path can just be directly archived without
+  having to look in the parent directory for transaction log files.
+
+  Wallet paths that point to files rather than directories will continue to be
+  accepted and interpreted the same as before.
 
 Credits
 =======